requirements.txt
bokeh
flask
requests
pymdown-extensions
PyYAML
jinja2
